refactor: use new HTML API for Scalar
