+++ src/renderers/scalar.ts
@@ -1,4 +1,4 @@
-import type { ApiReferenceConfiguration as ScalarConfig } from "@scalar/api-reference";
+import { getHtmlDocument, type HtmlRenderingConfiguration as ScalarConfig } from '@scalar/core/libs/html-rendering'
 import type { RenderHTMLOptions } from "../types.ts";
 
 // https://github.com/scalar/scalar
@@ -10,28 +10,10 @@ export default function render(opts: RenderHTMLOptions): string {
 
   const scalarConfig: ScalarConfig = {
     ...opts.scalar,
+    pageTitle: opts.meta?.title || "OpenAPI Documentation",
+    cdn: CDN_URL,
     url: opts.spec,
-    // @ts-expect-error (missing types?)
-    spec: { url: opts.specURL, ...opts.scalar?.spec },
   };
 
-  return /* html */ `<!doctype html>
-    <html lang="en">
-      <head>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <meta name="description" content="${opts.meta?.description}" />
-        <title>${opts.meta?.title}</title>
-        <style>${opts.styles}</style>
-      </head>
-      <body>
-        <script
-          id="api-reference"
-          data-configuration="${JSON.stringify(scalarConfig)
-            .split('"')
-            .join("&quot;")}"
-        ></script>
-        <script src="${CDN_URL}"></script>
-      </body>
-    </html>`;
+  return /* html */ getHtmlDocument(scalarConfig, opts.styles);
 }
